requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.1
boto3==1.26.142
botocore==1.29.142
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
clean-text==0.4.0
cleantext==1.1.4
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
Django==3.2
django-cors-headers==3.14.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
dynaconf==3.1.12
emoji==1.7.0
ftfy==6.1.1
hvac==1.1.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
knox==0.1.14
loguru==0.7.0
lxml==4.9.2
MarkupSafe==2.1.2
nltk==3.8.1
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
s3transfer==0.6.1
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2023.3
unicode==2.9
Unidecode==1.3.6
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
win32-setctime==1.1.0
